Updated as per latest changes in rust branch
