+++ bin/spiced/src/lib.rs
@@ -9,14 +9,11 @@ use app::App;
 use clap::Parser;
 use flightrepl::ReplConfig;
 use runtime::config::Config as RuntimeConfig;
-use runtime::dataconnector::DataConnector;
-use runtime::datafusion::DataFusion;
 use runtime::model::Model;
 
 use runtime::podswatcher::PodsWatcher;
 use runtime::Runtime;
 use snafu::prelude::*;
-use spicepod::component::dataset::{Dataset, Mode};
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -78,7 +75,11 @@ pub async fn run(args: Args) -> Result<()> {
 
     let mut df = runtime::datafusion::DataFusion::new();
 
-    load_datasets(&app, &mut df, &auth).await?;
+    for ds in &app.datasets {
+        Runtime::load_dataset(ds, &mut df, &auth)
+            .await
+            .context(UnableToLoadDatasetSnafu)?;
+    }
 
     let model_map = load_models(&app, &auth);
 
@@ -87,6 +88,7 @@ pub async fn run(args: Args) -> Result<()> {
     let mut rt: Runtime = Runtime::new(args.runtime, app, df, model_map, pods_watcher);
 
     rt.start_pods_watcher()
+        .await
         .context(UnableToInitializePodsWatcherSnafu)?;
 
     rt.start_servers()
@@ -127,123 +129,4 @@ fn load_models(app: &App, auth: &runtime::auth::AuthProviders) -> HashMap<String
     }
 
     model_map
-}
-
-async fn load_datasets(
-    app: &App,
-    df: &mut DataFusion,
-    auth: &runtime::auth::AuthProviders,
-) -> Result<()> {
-    for ds in &app.datasets {
-        let ds = Arc::new(ds.clone());
-        if ds.acceleration.is_none() && !ds.is_view() {
-            tracing::warn!("No acceleration specified for dataset: {}", ds.name);
-            continue;
-        };
-
-        let source = ds.source();
-        let source = source.as_str();
-        let params = Arc::new(ds.params.clone());
-        let data_connector: Option<Box<dyn DataConnector>> =
-            get_dataconnector_from_source(source, auth, Arc::clone(&params)).await?;
-
-        initialize_dataconnector(data_connector, df, source, &ds).await?;
-
-        tracing::info!("Loaded dataset: {}", &ds.name);
-    }
-
-    Ok(())
-}
-
-async fn get_dataconnector_from_source(
-    source: &str,
-    auth: &runtime::auth::AuthProviders,
-    params: Arc<Option<HashMap<String, String>>>,
-) -> Result<Option<Box<dyn DataConnector>>> {
-    match source {
-        "spice.ai" => Ok(Some(Box::new(
-            dataconnector::spiceai::SpiceAI::new(auth.get(source), params)
-                .await
-                .context(UnableToInitializeDataConnectorSnafu {
-                    data_connector: source,
-                })?,
-        ))),
-        "dremio" => Ok(Some(Box::new(
-            dataconnector::dremio::Dremio::new(auth.get(source), params)
-                .await
-                .context(UnableToInitializeDataConnectorSnafu {
-                    data_connector: source,
-                })?,
-        ))),
-        "localhost" | "" => Ok(None),
-        "debug" => Ok(Some(Box::new(dataconnector::debug::DebugSource {}))),
-        _ => UnknownDataConnectorSnafu {
-            data_connector: source,
-        }
-        .fail()?,
-    }
-}
-
-async fn initialize_dataconnector(
-    data_connector: Option<Box<dyn DataConnector>>,
-    df: &mut DataFusion,
-    source: &str,
-    ds: &Arc<Dataset>,
-) -> Result<()> {
-    let view_sql = ds.view_sql().context(InvalidSQLViewSnafu)?;
-
-    match data_connector {
-        Some(data_connector) => {
-            let data_backend = df
-                .new_accelerated_backend(Arc::clone(ds))
-                .context(UnableToCreateBackendSnafu)?;
-            let data_backend = Arc::new(data_backend);
-
-            if ds.mode() == Mode::ReadWrite {
-                if let Some(data_publisher) = data_connector.get_data_publisher() {
-                    df.attach_publisher(&ds.name.clone(), Arc::clone(ds), Arc::new(data_publisher))
-                        .await
-                        .context(UnableToAttachDataConnectorSnafu {
-                            data_connector: source,
-                        })?;
-                } else {
-                    tracing::warn!(
-                        "Data connector {source} does not support writes, but dataset {ds_name} is in read-write mode",
-                        ds_name = ds.name
-                    );
-                }
-
-                df.attach_publisher(&ds.name.clone(), Arc::clone(ds), Arc::clone(&data_backend))
-                    .await
-                    .context(UnableToAttachDataConnectorSnafu {
-                        data_connector: source,
-                    })?;
-            }
-
-            df.attach_connector_to_publisher(
-                Arc::clone(ds),
-                data_connector,
-                Arc::clone(&data_backend),
-            )
-            .context(UnableToAttachDataConnectorSnafu {
-                data_connector: source,
-            })?;
-        }
-        None => {
-            if view_sql.is_some() {
-                df.attach_view(ds).context(UnableToAttachViewSnafu)?;
-            } else {
-                let data_backend = df
-                    .new_accelerated_backend(Arc::clone(ds))
-                    .context(UnableToCreateBackendSnafu)?;
-                df.attach_publisher(&ds.name.clone(), Arc::clone(ds), Arc::new(data_backend))
-                    .await
-                    .context(UnableToAttachDataConnectorSnafu {
-                        data_connector: source,
-                    })?;
-            }
-        }
-    }
-
-    Ok(())
 }
